\name{had}
\alias{had100}
\alias{had116}
\alias{had156}
\alias{had172}
\alias{had188}
\alias{had236}
\alias{had244}
\alias{had52}
\alias{had92}

\docType{data}
\title{
Stored Hadmard matrices
}
\description{
Some stored Hadmard matrices, used in \code{hadamard.matrix}
}
\details{
Stored matrices from http://www2.research.att.com/~njas/hadamard/ 
filling the gaps up to 256 in \code{hadamard.matrix}, 260 is the next gap.
}
%%\usage{
%%data(had100)
%%data(had116)
%%data(had156)
%%data(had172)
%%data(had188)
%%data(had236)
%%data(had244)
%%data(had52)
%%data(had92)
%%}

\keyword{datasets}
