\name{hemp}
\alias{hemp}
\docType{data}
\title{
Hemp data
}
\description{
age and height of hemp plants.
}
\usage{data(hemp)}
\format{
  A data frame with 14 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\references{
Dieter Rasch, Juergen Pilz, L.R. Verdooren, Albrecht Gebhardt: Optimal Experimental Design with R, Chapman and Hall/CRC, 2011
}
\keyword{datasets}
