\name{OPDOE-undocumented}
\title{(still) undocumented functions}

\alias{add_size.mean.two_stage}
\alias{beta}
\alias{conf.matrix}
\alias{Confint.mean.two_stage}
\alias{conversion.low_high.two_levels}
\alias{decide.triangular.test}
\alias{defining_relation.fractional_factorial.two_levels}
\alias{delta.t.test}
\alias{hadamard.matrix}
\alias{hadamard.table}
\alias{is.hadamard}
\alias{jacobsthal}
\alias{legendre.symbol}
\alias{ncp}
\alias{optim_size_n.three_way_nested.model_4_a}
\alias{optim_size_n.three_way_nested.model_4_c}
\alias{paley}
\alias{paley2}
\alias{plot.triangular.test}
\alias{seqadd.confint}
\alias{seqadd.normal}
\alias{size.comparing.variances}
\alias{size.comparing_probs.mcnemar}
\alias{size.multiple_t.test}
\alias{size.prop.confint}
\alias{size.prop.test}
\alias{size.prop_test.two_sample}
\alias{size.selection.bechhofer}
\alias{size.seq_select.mean}
\alias{size.t.test}
\alias{size.variance.confint}
\alias{size_n.regII.ci_b1}
\alias{size_n.regII.ci_exp}
\alias{size_n.regII.ci_rho}
\alias{size_n.regII.test_rho}
\alias{size_n.regII.test_rho2}
\alias{size_n.regII.test_rho_2}
\alias{size_n.three_way_nested.model_4_c}
\alias{sizes.confint.welch}
\alias{sylvester}
\alias{walsh}
\alias{bibd}
\alias{print.bibd}
\alias{myGF}
\alias{myredu.modp}

\description{
  Undocumented / internal functions 
}
\references{
Dieter Rasch, Juergen Pilz, L.R. Verdooren, Albrecht Gebhardt: Optimal Experimental Design with R, Chapman and Hall/CRC, 2011  
}
\details{
  Some of these functions are not intended to be called by the user,
  others still lack their own documentation page. In the mean time see
  the referenced book.
 
}
\keyword{doe}
