\name{print.design.regression}
\alias{print.design.regression}
\title{
Prints a regression design object
}
\description{
Print method for a \code{\link{design.regression}} object.
}
\usage{
\method{print}{design.regression}(x, epl = 6, ...)
}
\arguments{
  \item{x}{
    \code{\link{design.regression}} object
}
  \item{epl}{
    integer, entries per line
}
  \item{\dots}{
    additional print arguments
}
}
\references{
  Dieter Rasch, Juergen Pilz, L.R. Verdooren, Albrecht Gebhardt:
  Optimal Experimental Design with R, Chapman and Hall/CRC, 2011
}   
\author{
  Dieter Rasch, Juergen Pilz, L.R. Verdooren, Albrecht Gebhardt
}
\keyword{lm}
\seealso{
\code{\link{design.regression}}
}
