\name{print.triangular.test}
\alias{print.triangular.test}
\alias{print.triangular.test.norm}
\alias{print.triangular.test.prop}
\title{Print method for Triangular Test Objects}
\usage{
\method{print}{triangular.test}(x, ...)
}
\arguments{
  \item{x}{\code{\link{triangular.test}} object}
  \item{...}{additional paramters for \code{print}}
}
\description{
  Prints a \code{triangular.test} object.
}
\seealso{\code{\link{triangular.test.norm}}, \code{\link{triangular.test.prop}}
}

\references{
  Dieter Rasch, Juergen Pilz, L.R. Verdooren, Albrecht Gebhardt:
  Optimal Experimental Design with R, Chapman and Hall/CRC, 2011
}   
\author{
  Dieter Rasch, Juergen Pilz, L.R. Verdooren, Albrecht Gebhardt
}
\keyword{test}
